feat: Renaming and attempting to iron out a bug with quote processing
