+++ src/utils.rs
@@ -31,6 +31,7 @@ pub fn read_lines(fname: &String) -> Result<Vec<String>, Error> {
 /// Function to write to a newly created file.
 pub fn write_to_file(name: String, finished: String) -> Option<()> {
     println!("Writing Snippets to file: {}", name);
+    println!("{}", finished);
     let mut outfile = File::create(Path::new("./").join(name))
         .unwrap_or_else(|err| panic!("Could not create the file {}", err));
